:recycle: refactor: 냉장고 공유 관련 api 수정(#53)
