+++ src/main/java/com/sm/project/web/controller/member/MemberController.java
@@ -22,7 +22,6 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -30,7 +29,6 @@ import javax.mail.MessagingException;
 import javax.validation.Valid;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
-import java.util.List;
 
 /**
  * MemberController는 회원 관련 API 요청을 처리하는 컨트롤러 클래스입니다.
@@ -251,52 +249,5 @@ public class MemberController {
         return ResponseDTO.onSuccess("패밀리 등록 성공");
     }
 
-    /**
-     * 냉장고 공유를 위해 사용자 추가 API입니다.
-     * @param request 이메일, 냉장고 Id
-     * @return 성공 메세지 반환합니다.
-     */
-    @PostMapping("/share")
-    @Operation(summary = "냉장고 공유 API", description = "이메일과 냉장고 Id를 입력하면 해당 이메일의 사용자를 냉장고에 등록하는 API입니다.")
-    public ResponseDTO<?> shareRefrigerator(Authentication authentication,
-            @RequestBody @Valid MemberRequestDTO.ShareDTO request){
-
-        Member member = memberQueryService.findMemberById(Long.valueOf(authentication.getName().toString())).orElseThrow(() -> new MemberHandler(ErrorStatus.MEMBER_NOT_FOUND));
-        memberService.shareRefrigerator(request);
-        return ResponseDTO.onSuccess("냉장고 사용자 추가 성공입니다.");
-    }
-
-    /**
-     * 냉장고에 등록된 사용자 조회 API입니다.
-     * @param authentication 인증 정보
-     * @param refrigeratorId 냉장고 id
-     * @return 사용자 정보 반홥
-     */
-    @GetMapping("/share/{refrigeratorId}")
-    @Operation(summary = "냉장고 등록된 사용자 조회 API", description = "냉장고 Id를 입력하면 등록된 사용자를 조회하는 API입니다.")
-    public ResponseDTO<?> getShareRefrigerator(Authentication authentication,
-                                               @PathVariable(name = "refrigeratorId") Long refrigeratorId){
-
-        List<Member> memberList = memberService.getShare(refrigeratorId);
-
-        return ResponseDTO.of(SuccessStatus.SHARE_GET_SUCCESS, MemberConverter.toShare(memberList));
-    }
-
-    /**
-     * 냉장고에 등록된 사용자 삭제 API입니다.
-     * @param authentication 인증 정보
-     * @param memberId 사용자 id
-     *
-     * @return 성공 메세지 반환
-     */
-    @DeleteMapping("/share/{memberId}")
-    @Operation(summary = "냉장고에 등록된 사용자 삭제 API", description = "사용자 id를 입력하면 냉장고를 공유한 사용자가 삭제됩니다.")
-    public ResponseDTO<?> deleteShareRefrigerator(Authentication authentication,
-                                                  @PathVariable(name = "memberId") Long memberId){
-
-        memberService.deleteShare(memberId);
-
-        return ResponseDTO.onSuccess("공유된 사용자 삭제 성공");
-    }
 
 }
